Move the historyMem test into a non-js test file
historyMem exists only in the non-js build; go vet for GOOS=js compiles
test files too and failed on the reference.
